Remove duplicates when two pattern matches the same event
If a community has patterns configured like this :

```
"patternsGoogleCalendar": ["Apéro PHP", "AFUP"],
```

and if an event has a title "Apéro PHP" and a link
in the description that contains "afup" (for exemple
a link to "lyon.afup.org",

then then event is displayed twice (because the event
is found is both requests to google calendar).

To avoid that we remove duplicates from the event list,
based on the google calendar id.